% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_time_mgmt.r
\name{as_timestamptz_vec}
\alias{as_timestamptz_vec}
\title{Vectorized assigning of timezones, for (potentially) multiple timezones}
\usage{
as_timestamptz_vec(date, time, timezone, tz_out = NA)
}
\description{
Takes in date, time, and timezone vectors, and combines them into a single timestamp with timezone vector.
}
