% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_maintenance.r
\name{catagorical_check}
\alias{catagorical_check}
\title{Check for valid values in catagorical variables}
\usage{
catagorical_check(data, table, mdc)
}
\arguments{
\item{data}{Data table with column names matching a given table in the column metadata}

\item{table}{corresponding table name in column metadata (str)}

\item{novel_data}{column metadata pulled from database}
}
\value{
Silently returns same data table with no revisions. Throws warning with report on invalid data if found.
}
\description{
Compares caagorical varibles with the defined catagories in column metadata, to make sure they are validprior to pushing.
}
