% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.r
\name{check_package_version}
\alias{check_package_version}
\title{Check package version against GitHub repository}
\usage{
check_package_version()
}
\description{
Check package version against GitHub repository
}
\keyword{internal}
