% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_time_mgmt.r
\name{clean_survey_times}
\alias{clean_survey_times}
\title{Clean up survey times}
\usage{
clean_survey_times(df, tz)
}
\arguments{
\item{df}{dataframe with "date", "start_time", "end_time" columns}

\item{tz}{valid timezone as sting from OlsonNames()}
}
\description{
takes in dataframe with "date", "start_time", "end_time" columns, as well as a local timezone tz, calculates start_timestamp_utc, end_timestamp_utc, and duration_minutes.
}
