% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_maintenance.r
\name{compare_updates}
\alias{compare_updates}
\title{Distinguishing columns to update}
\usage{
compare_updates(cfs_results, id_cols = NA)
}
\arguments{
\item{data_old}{Results from \link[ribbitrrr]{compare_for_staging} which contain updates (either update = TRUE or return_all = TRUE)}
}
\value{
A list of dataframes corresponding to columns which differe between update and update_old data.
}
\description{
Using output results from \link[ribbitrrr]{compare_for_staging}, identifies which columns differ for a given set of rows and outputs a comparison table for each column.
}
\examples{

data_old = data.frame(id= 1:5, values = c("a", "b", "c", "d", "e"))
data_new = data.frame(id = 3:7, values = c("d", "d", "f", "f", "g"))

comparison <- compare_for_staging(data_old, data_new, key_columns = c("id"), return_all = TRUE)
update_diffs <- compare_updates(comparison)
}
