% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_time_mgmt.r
\name{count_observers}
\alias{count_observers}
\title{Count observers}
\usage{
count_observers(observers)
}
\arguments{
\item{start_timestamp}{timestamp #1}

\item{end_timestamp}{timestamp #2}
}
\description{
counts number of observers in sting (separated by ","). use purrr::map_chr to apply within groups.
}
