% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_time_mgmt.r
\name{duration_minutes}
\alias{duration_minutes}
\title{time duration from timestamp}
\usage{
duration_minutes(start_timestamp, end_timestamp)
}
\arguments{
\item{start_timestamp}{timestamp #1}

\item{end_timestamp}{timestamp #2}
}
\description{
calculates time duration from start and end timestamps in minutes, rounded to the nearest whole
}
