% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_time_mgmt.r
\name{find_time_range}
\alias{find_time_range}
\title{Determine first and last times from times without dates, assuming less than 24 hours among group.}
\usage{
find_time_range(times, max_range = 12)
}
\arguments{
\item{times}{vector of all times to be considered}

\item{max_range}{max range considered when calculating times, in hours. must be < 24}
}
\description{
Takes in times, determines the first and last values assuming 24h time format and a max time range "max_range" < 24h
}
