% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_metadata.r
\name{get_query_metadata}
\alias{get_query_metadata}
\title{Extract column metadata from a dbplyr lazy table object}
\usage{
get_query_metadata(dbcon, lazy_tbl)
}
\arguments{
\item{dbcon}{DBI database connection object}

\item{lazy_tbl}{A dbplyr lazy table object}
}
\value{
A dataframe with schema_name, table_name, and column_name for each output column
}
\description{
Extract column metadata from a dbplyr lazy table object
}
