% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_access.r
\name{hopRegister}
\alias{hopRegister}
\title{hopRegister}
\usage{
hopRegister(dbcon, dbhost, dbname = NA)
}
\arguments{
\item{dbcon}{DBI connection object}

\item{dbhost}{database host}

\item{dbname}{an identifier (string) for the connection}
}
\description{
Register connection in RStudio Connections Pane
}
