% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_access.r
\name{HopToDB}
\alias{HopToDB}
\title{HopToDB}
\usage{
HopToDB(db_credentials, timezone = NA)
}
\arguments{
\item{db_credentials}{A list containing the following database connection credentials (see \link[DBI]{dbConnect}): dbname - the database name, host - the database host, port - the database port, user - your database username, password - your database password.}

\item{timezone}{an optional timezone parameter to help convert data from various timezones to your local time. Valid options are found using \link[lubridate]{OlsonNames}.}
}
\value{
database connection object, to be passed to other functions (e.g. \link[DBI]{dbListTables}, \link[dbplyr]{tbl}, etc.).
}
\description{
A simple function to easily to the RIBBiTR (or another) remote database.
}
\examples{

# open your local .Renviron filero
usethis::edit_r_environ()

# copy the following to .Renviron, replacing corresponding database credentials
ribbitr_db = list(
  dbname = "[DATABASE_NAME]"
  host = "[DATABASE_HOST]"
  port = "[DATABASE_PORT]"
  user = "[USERNAME]"
  password = "[PASSWORD]"
)

# connect to your database with a single line of code
dbcon <- HopToDB(ribbitr_db)

}
