% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microclimate.r
\name{local_time_vec}
\alias{local_time_vec}
\title{Convert timestamp to local time}
\usage{
local_time_vec(timestamptz, local_tz, drop_tz = TRUE)
}
\arguments{
\item{timestamptz}{Column name of timestamp data}

\item{local_tz}{Column name specifying the local timezone, otherwise a valid timezone (sting) \link[base]{OlsonNames}}

\item{drop_tz}{do you want to discard timezone info in output? If yes, returns a datetime column. If no, returns a list of datetimes with (possibly) varying timezone attached.}
}
\value{
Returns data frame with additional column specified
}
\description{
Takes in timestamp with timezone data and local timezone, outputs in local time.
}
\examples{

data_local = data \%>\%
mutate(local_time_data = local_time_db(timestamptz, tz))

}
