% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_microclimate.R
\name{microclimate_presurvey}
\alias{microclimate_presurvey}
\title{Query microclimate data for a buffer period preceding a given survey}
\usage{
microclimate_presurvey(
  survey_data,
  dbcon,
  buffer,
  time_series_pointer,
  time_sensitive = FALSE,
  output_timezone = NA
)
}
\arguments{
\item{survey_data}{table containing columns survey_id, site_id, date, start_time (or start_timestamp_utc if time_sensitive = TRUE)}

\item{dbcon}{DBI database connection object (e.g. returned by \link[ribbitrrr]{hopToDB})}

\item{buffer}{a lubridate time period (e.g. \link[lubridate]{days}) of interest for time series data query preceding each survey}

\item{time_series_pointer}{a \link[dplyr]{tbl} object for the time series table of interest (e.g. tbl(dbcon, Id("microclimate_data", "ts_temperature") for temperature data)}

\item{time_sensitive}{Boolean: Do you want to query by time using survey start_timestamp_utc? If FALSE, queried by date only.}

\item{output_timezone}{What timezone do you want resulting data to be output in (default: NA <=> UTC)? Format as string, options found in \link[base]{OlsonNames}}
}
\value{
coinciding time series data (where available) with an associated survey_id
}
\description{
From a table of surveys, returns microclimate data from desired timeseries from the available, corresponding sites for a buffer window preceing the survey start.
}
