% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_maintenance.r
\name{resolve_sample_conflicts}
\alias{resolve_sample_conflicts}
\title{Identify and resolve sample name conflicts}
\usage{
resolve_sample_conflicts(data, db_sample)
}
\arguments{
\item{data}{Cleaned new sample data to be compared for conflicts}

\item{db_sample}{A lazy table object pointing to the sample database table ("old data")}

\item{novel_data}{Data Frame of rows to be pushed to the database}
}
\value{
Returns a table of new data plus any old data to be updated. This table should be upserted into database.
}
\description{
Identifies sample_name/sample_type conflicts 1) within new data, 2) between new and old data, and 3) between new and old, previously conflicting and resolved data.
}
