% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy.r
\name{ribbitr_clean_taxa}
\alias{ribbitr_clean_taxa}
\title{Systematically clean taxa names}
\usage{
ribbitr_clean_taxa(data, taxon_column, comment_column = NULL)
}
\arguments{
\item{data}{data frame with taxa data}

\item{taxon_column}{column with taxon names}

\item{comment_column}{optional column, for remapping some taxon details to comments prior to cleaning}
}
\value{
Data frame with cleaned taxa
}
\description{
References a dictionary to remap taxa names, adding comments where needed.
}
\examples{
# example data
my_data <- tibble(
  taxa_names = c("desmog_spp", "rana_muscosa", "pristimantis_sp._potential_new_sp", "ololygon_aff._littoralis"),
  comments = c(NA, "very cute", NA, "got away"),
  other_data = c(1, 3, 5, 7)
)
clean_data = my_data \%>\%
  ribbitr_clean_taxa(taxa_names, comments)
}
