% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy.r
\name{ribbitr_taxa_lookup}
\alias{ribbitr_taxa_lookup}
\title{Systematically look up taxa names}
\usage{
ribbitr_taxa_lookup(
  taxa,
  itis = TRUE,
  ncbi = TRUE,
  gbif = TRUE,
  iucn = TRUE,
  cites = FALSE,
  cites_token = NA,
  output = "full"
)
}
\arguments{
\item{taxa}{data frame with taxa data}

\item{itis}{do you want to query ITIS?}

\item{ncbi}{do you want to query NCBI?}

\item{gbif}{do you want to query GBIF?}

\item{iucn}{do you want to query IUCN?}

\item{cites}{do you want to query CITES?}

\item{cites_token}{CITES API token, required to query CITES.}

\item{output}{how do you want the output formatted: "full" returns everything; "simplified" returns less, "simple" returns even less.}
}
\value{
table of results for submitted taxa
}
\description{
References a number of taxanomic databases to build or update a taxonomy lookup table. Primary reference in Amphibiaweb, others are optional.
}
\examples{
# example data

my_taxa = c(
"rana_muscosa",
"atelopus_zeteki",
"pristimantis")

my_taxa_lookup = ribbitr_taxa_lookup(my_taxa)
}
