% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_maintenance.r
\name{stage_to_temp}
\alias{stage_to_temp}
\title{Stage data to temporary database table}
\usage{
stage_to_temp(dbcon, reference_table, novel_data)
}
\arguments{
\item{dbcon}{Database connection (A valid and active DBI database connection object)}

\item{reference_table}{A lazy table object pointing to the database table which you want to push data to.}

\item{novel_data}{Data Frame of rows to be pushed to the database}
}
\value{
If successful transaction, the name of the temporary table holding the data.
}
\description{
This function allows you to push new data to a temporary table of the same constraints/datatypes as the table you want to eventually push to, to make sure everything checks out. Once data is staged, it can be pushed (inserted, updated, upserted) to the corresponding database table using ____.
}
\examples{
db_capture = tbl(dbcon, "capture")
temp_capture_tbl = stage_to_temp(dbcon, db_capture, novel_capture)
pointer = tbl(dbcon, temp_capture_tbl)
rows_upsert(db_capture, pointer, by=capture_pkey, in_place=TRUE)
}
