% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_time_mgmt.r
\name{start_timestamp_local}
\alias{start_timestamp_local}
\title{start_timestamp_local}
\usage{
start_timestamp_local(date, start_time, tz)
}
\arguments{
\item{date}{date as Date}

\item{start_time}{as time}

\item{tz}{timezone from OlsonNames()}
}
\description{
converts time, date, timezone to a timestamp
}
