% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_access.r
\name{tbl_chain}
\alias{tbl_chain}
\title{Recursively identify commutative reference ("explicitly or implicitly linked") tables}
\usage{
tbl_chain(tbl_name, metadata_columns, until = NA)
}
\arguments{
\item{tbl_name}{The name of the table of interest (string)}

\item{metadata_columns}{Column metadata containing the table of interest (Data Frame)}

\item{until}{Name(s) of table(s) beyond which you do not want to continue link search (string or list of strings)}
}
\value{
Returns a link object (list of all encountered referenced tables and their attributes)
}
\description{
For each foreign key in tbl_name, pull information on the associated reference table. Do the same for any foreign keys found in this reference table, and so on. Stops when table network is exausted. Does not search beyond any tables listed in "until".
}
\examples{

capture_chain = tbl_chain("capture", mdc, until=c("region"))
}
