% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_access.r
\name{tbl_fkey}
\alias{tbl_fkey}
\title{Table foreign key}
\usage{
tbl_fkey(tbl_name, metadata_columns)
}
\arguments{
\item{tbl_name}{The name of the table of interest (string)}

\item{metadata_columns}{Column metadata containing the table of interest (Data Frame)}
}
\value{
The name(s) of the columns with foreign key status in the table provided
}
\description{
Identify the foreign key column(s) for a given table in the database metadata. A foreign key points to and mirrors the primary key of another table, establishing relationships between tables. There may be multiple foreign keys in a given table, usually (though not necessarily) single column per key.
}
\examples{
survey_fkey <- tbl_fkey('survey', mdc)
}
