% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_access.r
\name{tbl_fkey_ref}
\alias{tbl_fkey_ref}
\title{Table foreign key reference columns}
\usage{
tbl_fkey_ref(tbl_schema, tbl_name, tbl_fkey_col, metadata_columns)
}
\arguments{
\item{tbl_schema}{The name of the schema of the table of interest (string)}

\item{tbl_name}{The name of the table of interest (string)}

\item{tbl_fkey_col}{The name of the fkey column in the table of interest (string)}

\item{metadata_columns}{Column metadata containing the table of interest (Data Frame)}
}
\value{
The name(s) of the columns with foreign key status in the table provided
}
\description{
Identify the foreign key reference column for a given foreign key.
}
\examples{
dbcon <- hopToDB("ribbitr")

mdc <- tbl(dbcon, Id("public", "all_columns")) \%>\%
            filter(table_schema =="survey_data") \%>\%
            collect()

survey_fkey <- tbl_fkey('survey', mdc)
}
