% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_access.r
\name{tbl_join}
\alias{tbl_join}
\title{Join table with reference tables}
\usage{
tbl_join(dbcon, link, tbl = NA, join = "left", by = "pkey", columns = NA)
}
\arguments{
\item{dbcon}{database connection object from \link[DBI]{dbConnect} or \link[ribbitrrr]{HopToDB}}

\item{link}{a link object generated from \link[ribbitrrr]{tbl_link} or \link[ribbitrrr]{tbl_chain}}

\item{tbl}{A lazy table object from \link[dplyr]{tbl} corresponding to the root table in the provided link object (optional). If not provided, link object root table will be pulled in its entirity. Passing your own table allows you to filter or select specific columns prior to joining, thereby avoiding pulling nonessential data. Be sure to minimally include essential columns: fkeys, and nkeys or pkeys depending on "by".}

\item{join}{Type of join to be executed ("left", "inner", "full", or "right")}

\item{by}{What columns to perform join on ("pkey" or "nkey")}

\item{columns}{Additional columns to be included from joined tables (string or list of stings). Primary, natural, and foreign key columns are included by default.}
}
\value{
Returns a single lazy table object of all linked tables joined as specified
}
\description{
Recursively join linked database tables following provided link object. Only primary, natural, and foreign key columns are joined by default. Specify additional columns to include in "columns"
}
\examples{
# generate link object
capture_chain = tbl_chain("capture", mdc, until=c("region"))

# pre-filter root table (optional)
tbl_capture = tbl(dbcon, Id("survey_data", "capture")) \%>\%
  select(all_of(tbl_keys("capture", mdc)),
       species_capture,
       bd_swab_id) \%>\%
  filter(!is.na(bd_swab_id))

# create and filter join table
tbl_capture_brazil = tbl_join(dbcon, capture_chain, tbl=db_capture) \%>\%
  filter(region == "brazil")

# collect (pull) data from database
capture_brazil = tbl_capture_brazil \%>\%
  collect()

}
