% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_access.r
\name{tbl_keys}
\alias{tbl_keys}
\title{All table keys}
\usage{
tbl_keys(tbl_name, metadata_columns)
}
\arguments{
\item{tbl_name}{The name of the table of interest (string)}

\item{metadata_columns}{Column metadata containing the table of interest (Data Frame)}
}
\value{
The uniqie name(s) of the columns with primary, foreign, or natural key status in the table provided. If a column has multiple statuses, it will show up only once.
}
\description{
Identify primary, foreign, and natural key columns for a given table.
}
\examples{
survey_keys <- tbl_keys('survey', mdc)
db_survey <- tbl(dbcon, "survey") \%>\%
                select(all_of(survey_keys)) \%>\%
                collect()
}
