% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_access.r
\name{tbl_link}
\alias{tbl_link}
\title{Identify reference ("explicitly linked") tables}
\usage{
tbl_link(tbl_name, metadata_columns)
}
\arguments{
\item{tbl_name}{The name of the table of interest (string)}

\item{metadata_columns}{Column metadata containing the table of interest (Data Frame)}
}
\value{
Returns a link object listing referenced tables and their attributes
}
\description{
For each foreign key in tbl_name, pull information on the associated reference table.
}
\examples{

capture_link = tbl_link("capture", mdc)
}
