% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_access.r
\name{tbl_nkey}
\alias{tbl_nkey}
\title{Table natural key}
\usage{
tbl_nkey(tbl_name, metadata_columns)
}
\arguments{
\item{tbl_name}{The name of the table of interest (string)}

\item{metadata_columns}{Column metadata containing the table of interest (Data Frame)}
}
\value{
The name(s) of the columns with natural key status in the table provided
}
\description{
Identify the natural key column(s) for a given table in the database metadata.  A natural key is comprised of the columns which naturally identify a given row, and which are generally used to generate ID columns used as primary keys. There is only one natural key per table, though it may be comprised of multiple columns (a "composite" key).
}
\examples{
survey_nkey <- tbl_nkey('survey', mdc)
}
