% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_exploration.r
\name{tbl_pkey}
\alias{tbl_pkey}
\title{Table primary key}
\usage{
tbl_pkey(tbl_name, metadata_columns)
}
\arguments{
\item{tbl_name}{The name of the table of interest (string)}

\item{metadata_columns}{Column metadata containing the table of interest (Data Frame)}
}
\value{
The name(s) of the columns comprising the primary key for the table provided.
}
\description{
Identify the primary key column(s) for a given table in the database metadata. A table's primary key caries a unique and not-null constraint, and is used to uniquely identify the rows in the table. There is only one primary key per table, usually (though not necessarily) single column.
}
\examples{
survey_pkey <- tbl_pkey('survey', mdc)
}
