% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_access.r
\name{tbl_right_join}
\alias{tbl_right_join}
\title{Right join using link object}
\usage{
tbl_right_join(dbcon, link, tbl = NA, by = "pkey", columns = "all")
}
\arguments{
\item{dbcon}{database connection object from \link[DBI]{dbConnect} or \link[ribbitrrr]{hopToDB}}

\item{link}{a link object generated from \link[ribbitrrr]{tbl_link} or \link[ribbitrrr]{tbl_chain}}

\item{tbl}{A lazy table object from \link[dplyr]{tbl} corresponding to the root table in the provided link object (optional). If not provided, link object root table will be pulled in its entirity. Passing your own table allows you to filter or select specific columns prior to joining, thereby avoiding pulling nonessential data. Be sure to minimally include essential columns: fkeys, and nkeys or pkeys depending on "by".}

\item{by}{What columns to perform join on ("pkey" or "nkey")}

\item{columns}{Additional columns to be included from joined tables (string or list of stings). Set to "all" to include all columns. Primary, natural, and foreign key columns are always included by default (columns = NA)}
}
\value{
Returns a single lazy table object of all linked tables joined as specified
}
\description{
Right join tables across a provided link object, generated using \link[ribbitrrr]{tbl_link} or \link[ribbitrrr]{tbl_chain}
}
\examples{
dbcon <- hopToDB("ribbitr")

mdc <- tbl(dbcon, Id("survey_data", "metadata_columns")) \%>\%
  filter(table_schema == "survey_data") \%>\%
  collect()

# generate link object
capture_chain = tbl_chain("capture", mdc)

# pre-filter root table (optional)
db_capture = tbl(dbcon, Id("survey_data", "capture")) \%>\%
  select(all_of(tbl_keys("capture", mdc)),
         species_capture,
         bd_swab_id) \%>\%
  filter(!is.na(bd_swab_id))

# create and filter join table
tbl_capture_brazil = tbl_right_join(dbcon, capture_chain, tbl=db_capture) \%>\%
  filter(location == "brazil")

# collect (pull) data from database
capture_brazil = tbl_capture_brazil \%>\%
  collect()
}
