% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_time_mgmt.r
\name{timestamp_of_capture_utc}
\alias{timestamp_of_capture_utc}
\title{timestamp_of_capture_utc}
\usage{
timestamp_of_capture_utc(df)
}
\arguments{
\item{df}{dataframe with "date", "time_of_capture", and "time_zone" columns}
}
\description{
calculates timestamp of capture from date, time_of_capture, and time_zone. If the timestamp of capture is before the start timestamp, it adds 24 hours to the timestamp of capture.
}
