% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_maintenance.r
\name{compare_for_staging}
\alias{compare_for_staging}
\title{Compare data for staging}
\usage{
compare_for_staging(
  data_old,
  data_new,
  key_columns,
  return_all = TRUE,
  insert = TRUE,
  update = TRUE,
  orphan = FALSE,
  duplicate = FALSE,
  report = FALSE
)
}
\arguments{
\item{data_old}{A reference data frame containing the existing data}

\item{data_new}{A data frame containing the new data to be compared with the existing data}

\item{key_columns}{A character vector of key columns used to distinguish "insert", "update", and "orphan" rows. Can be a primary key or natural key, depending on needs.}

\item{return_all}{Logical: do you want to perform all comparisons (insert, update, orphan, duplicate)? Overrides previous logical parmeters. False by default.}

\item{insert}{Logical: do you want to check for new rows to insert? TRUE by default.}

\item{update}{Logical: do you want to check for new rows to update? TRUE by default.}

\item{orphan}{Logical: do you want to check for orphan rows in the old data? FALSE by default.}

\item{duplicate}{Logical: do you want to check for duplicate rows? FALSE by default.}

\item{report}{String name for report if desired. If provided, function will announce the lengths of all provided outputs.}
}
\value{
A list of dataframes containing rows corresponding to each comparison (insert, update, orphan, duplicate)
}
\description{
Compare new data with a corresponding existing database table prior to writing, to identify rows in the new data with keys which are not yet found in the existing data ("insert"), rows which are identical in all shared columns to rows in the existing data ("duplicate"), rows with existing key combination but different non-key values in the existing data ("update"), and rows in the existing data with key combinations which do not exist in the new data ("orphan")
}
\examples{

data_old = data.frame(id= 1:5, values = c("a", "b", "c", "d", "e"))
data_new = data.frame(id = 3:7, values = c("d", "d", "f", "f", "g"))

comparison <- compare_for_staging(data_old, data_new, key_columns = c("id"), return_all = TRUE)
data_to_insert = comparison$insert
data_to_update = comparison$update
data_orphans = comparison$orphan
data_duplicates = comparison$duplicate
}
