% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_access.r
\name{hopToDB}
\alias{hopToDB}
\title{hopToDB}
\usage{
hopToDB(prefix = NA, timezone = NULL)
}
\arguments{
\item{prefix}{an optional prefix (string) added to the front of connection credential variables to distinguish between sets of credentials.}

\item{timezone}{an optional timezone parameter to help convert data from various timezones to your local time.}
}
\value{
database connection object, to be passed to other functions (e.g. \link[DBI]{dbListTables}, \link[dplyr]{tbl}, etc.).
}
\description{
A simple function to easily to the RIBBiTR (or another) remote database. The following database connection credentials must be saved to your local .Renviron file(see \link[DBI]{dbConnect}): dbname - the database name, host - the database host, port - the database port, user - your database username, password - your database password. Variables may be defined with an optional prefix seperated by and underscore (e.g. "ribbitr_dbname") to distinguish multiple connections.
}
\examples{

## open your local .Renviron file
# usethis::edit_r_environ
## copy the following to .Renviron, replacing corresponding database credentials

# ribbitr.dbname = "[DATABASE_NAME]"
# ribbitr.host = "[DATABASE_HOST]"
# ribbitr.port = "[DATABASE_PORT]"
# ribbitr.user = "[USERNAME]"
# ribbitr.password = "[PASSWORD]"

# connect to your database with a single line of code
dbcon <- hopToDB("ribbitr")
}
