% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amphibiaweb.R
\name{scrape_amphibiaweb}
\alias{scrape_amphibiaweb}
\title{Scrape taxa metadata from AmphibiaWeb}
\usage{
scrape_amphibiaweb(sci_nam, quietly = FALSE)
}
\arguments{
\item{sci_nam}{Scientific name(s) of species of interest}
}
\value{
Associated metadata (if any)
}
\description{
Searches scientific names for corresponding species in AmphibiaWeb
}
\examples{
scientific_names <- c("Rana muscosa", "Oedipina gracilis")
results <- scrape_amphibiaweb(scientific_names)
print(results)
}
