% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_access.r
\name{tbl_join}
\alias{tbl_join}
\alias{tbl_left_join}
\alias{tbl_inner_join}
\alias{tbl_full_join}
\alias{tbl_right_join}
\title{Join tables with link object}
\usage{
tbl_join(dbcon, link, tbl = NA, join = "left", by = "pkey", columns = "all")
}
\arguments{
\item{dbcon}{database connection object from \link[DBI]{dbConnect} or \link[ribbitrrr]{hopToDB}}

\item{link}{a link object generated from \link[ribbitrrr]{tbl_link} or \link[ribbitrrr]{tbl_chain}}

\item{tbl}{A lazy table object from \link[dplyr]{tbl} corresponding to the root table in the provided link object (optional). If not provided, link object root table will be pulled in its entirity. Passing your own table allows you to filter or select specific columns prior to joining, thereby avoiding pulling nonessential data. Be sure to minimally include essential columns: fkeys, and nkeys or pkeys depending on "by".}

\item{join}{Type of join to be executed ("left", "inner", "full", or "right")}

\item{by}{What columns to perform join on ("pkey" or "nkey")}

\item{columns}{Additional columns to be included from joined tables (string or list of stings). Set to "all" to include all columns. Primary, natural, and foreign key columns are always included by default (columns = NA)}
}
\value{
Returns a single lazy table object of all linked tables joined as specified
}
\description{
These functions provide a framework for automatically and recursively joining tables using link objects.
Link objects are generated using \link[ribbitrrr]{tbl_link} or \link[ribbitrrr]{tbl_chain}.
\itemize{
\item \code{tbl_left_join()}: Left join tables in link object.
\item \code{tbl_inner_join()}: Inner join tables in link object.
\item \code{tbl_full_join()}: Full join tables in link object.
\item \code{tbl_right_join()}: Right join tables in link object.
\item \code{tbl_join()}: Generic join tables in link object (pass join type as parameter)
}

Recursively join linked database tables following provided link object. Only primary, natural, and foreign key columns are joined by default. Specify additional columns to include in "columns"
}
\examples{
dbcon <- hopToDB("ribbitr")

mdc <- tbl(dbcon, Id("survey_data", "metadata_columns")) \%>\%
  filter(table_schema == "survey_data") \%>\%
  collect()

# generate link object
capture_chain = tbl_chain("capture", mdc)

# pre-filter root table (optional)
db_capture = tbl(dbcon, Id("survey_data", "capture")) \%>\%
  select(all_of(tbl_keys("capture", mdc)),
         species_capture,
         bd_swab_id) \%>\%
  filter(!is.na(bd_swab_id))

# create and filter join table
tbl_capture_brazil = tbl_join(dbcon, capture_chain, tbl=db_capture) \%>\%
  filter(location == "brazil")

# collect (pull) data from database
capture_brazil = tbl_capture_brazil \%>\%
  collect()

}
